Moved util.h to the fuse_core package
