+++ fuse_constraints/include/fuse_constraints/normal_prior_pose_3d_cost_functor.h
@@ -34,9 +34,10 @@
 #ifndef FUSE_CONSTRAINTS_NORMAL_PRIOR_POSE_3D_COST_FUNCTOR_H
 #define FUSE_CONSTRAINTS_NORMAL_PRIOR_POSE_3D_COST_FUNCTOR_H
 
-#include <fuse_constraints/util.h>
 #include <fuse_constraints/normal_prior_orientation_3d_cost_functor.h>
+
 #include <fuse_core/eigen.h>
+#include <fuse_core/util.h>
 
 #include <Eigen/Core>
