+++ fuse_constraints/src/normal_delta_orientation_2d.cpp
@@ -32,7 +32,8 @@
  *  POSSIBILITY OF SUCH DAMAGE.
  */
 #include <fuse_constraints/normal_delta_orientation_2d.h>
-#include <fuse_constraints/util.h>
+
+#include <fuse_core/util.h>
 
 
 namespace fuse_constraints
@@ -52,8 +53,7 @@ bool NormalDeltaOrientation2D::Evaluate(
   // The following lines should read as
   // r = A_ * ((x1 - x0) - b_);
   // The wrap function handles the 2_pi wrap around issue with rotations
-  double angle_diff = parameters[1][0] - parameters[0][0] - b_;
-  wrapAngle2D(angle_diff);
+  double angle_diff = fuse_core::wrapAngle2D(parameters[1][0] - parameters[0][0] - b_);
   residuals[0] = A_ * angle_diff;
   if (jacobians != NULL)
   {
