+++ fuse_core/include/fuse_core/util.h
@@ -31,15 +31,16 @@
  *  ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
  *  POSSIBILITY OF SUCH DAMAGE.
  */
-#ifndef FUSE_VARIABLES_UTIL_H
-#define FUSE_VARIABLES_UTIL_H
+#ifndef FUSE_CORE_UTIL_H
+#define FUSE_CORE_UTIL_H
 
 #include <ceres/jet.h>
+#include <Eigen/Core>
 
 #include <cmath>
 
 
-namespace fuse_variables
+namespace fuse_core
 {
 
 /**
@@ -47,8 +48,8 @@ namespace fuse_variables
  *
  * @param[in] w The quaternion real-valued component
  * @param[in] x The quaternion x-axis component
- * @param[in] y The quaternion y-axis component
- * @param[in] z The quaternion z-axis component
+ * @param[in] y The quaternion x-axis component
+ * @param[in] z The quaternion x-axis component
  * @return      The quaternion's Euler pitch angle component
  */
 template <typename T>
@@ -72,8 +73,8 @@ static inline T getPitch(const T w, const T x, const T y, const T z)
  *
  * @param[in] w The quaternion real-valued component
  * @param[in] x The quaternion x-axis component
- * @param[in] y The quaternion y-axis component
- * @param[in] z The quaternion z-axis component
+ * @param[in] y The quaternion x-axis component
+ * @param[in] z The quaternion x-axis component
  * @return      The quaternion's Euler pitch angle component
  */
 template <typename T>
@@ -90,8 +91,8 @@ static inline T getRoll(const T w, const T x, const T y, const T z)
  *
  * @param[in] w The quaternion real-valued component
  * @param[in] x The quaternion x-axis component
- * @param[in] y The quaternion y-axis component
- * @param[in] z The quaternion z-axis component
+ * @param[in] y The quaternion x-axis component
+ * @param[in] z The quaternion x-axis component
  * @return      The quaternion's Euler pitch angle component
  */
 template <typename T>
@@ -119,6 +120,12 @@ void wrapAngle2D(T& angle)
   angle -= TAU * ceres::floor((angle + PI) / TAU);
 }
 
+/**
+ * @brief Wrap a 2D angle to the standard (-Pi, +Pi] range.
+ *
+ * @param[in] angle Input angle to be wrapped to the (-Pi, +Pi] range.
+ * @return The equivalent wrapped angle
+ */
 template<typename T>
 T wrapAngle2D(const T& angle)
 {
@@ -134,7 +141,7 @@ T wrapAngle2D(const T& angle)
  * @return          The equivalent 2x2 rotation matrix
  */
 template <typename T>
-Eigen::Matrix<T, 2, 2, Eigen::RowMajor> RotationMatrix2D(const T angle)
+Eigen::Matrix<T, 2, 2, Eigen::RowMajor> rotationMatrix2D(const T angle)
 {
   const T cos_angle = ceres::cos(angle);
   const T sin_angle = ceres::sin(angle);
@@ -143,6 +150,6 @@ Eigen::Matrix<T, 2, 2, Eigen::RowMajor> RotationMatrix2D(const T angle)
   return rotation;
 }
 
-}  // namespace fuse_variables
+}  // namespace fuse_core
 
-#endif  // FUSE_VARIABLES_UTIL_H
+#endif  // FUSE_CORE_UTIL_H
